+++ test/src/functions.py
@@ -1,7 +1,7 @@
 import unittest, time, sys, os
 # Add the root directory to the sys.path
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-import functions as func
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../..')))
+import src.functions as func
 
 
 class TestFunctions(unittest.TestCase):
@@ -15,13 +15,13 @@ class TestFunctions(unittest.TestCase):
         self.assertEqual(func.validate_username('this usernameis 12 chars over'), "INVALID: LONG")
     
     def test_validate_username_4(self):
-        self.assertEqual(func.validate_username('yo   asdf'), "INVALID: DOUBLE SPACE" )
+        self.assertEqual(func.validate_username('yo   asdf'), "INVALID: DOUBLE SPACE")
 
     def test_validate_username_5(self):
         self.assertEqual(func.validate_username('oj'), "INVALID: BANNED NAME")
 
     
     
-
+    
 if __name__ == '__main__':
     unittest.main()
